Added the HashSet and LinkedList
Some notes:
the HashMap contains private properties and
methods, to make it more secure.

The HashSet doesnt contain that, and when i was
trying to extend the HashMap class into the
HashSet class, i was getting problems with the
private methods.
So i decided to implement it from scratch.

Also, both classes use the LinkedList and Node
classes from the previous exercise, thats why i
added them to the file.